Scion Setup Specification Option (#42)
SCION setup specification option

Allow users to specify which SCION stack they want to deploy
on nodes in the simulation. The official .deb packages
are supported, as well as local builds of custom repo checkouts.
This removes hard dependency on scion lab packages.
The setup specification is just an ordinary option whose value can be specified
and overridden per AS or even individual node.

Co-authored-by: Divine Paul <divineflu@gmail.com>